multi: move merkle to lnwire
